refactor(Controller): 뷰 컨트롤러 통합
