+++ src/main/java/roomescape/page/PageController.java
@@ -2,10 +2,19 @@ package roomescape.page;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
-public class AdminController {
+public class PageController {
+
+    @GetMapping("/")
+    public String getPopularThemePage() {
+        return "index";
+    }
+
+    @GetMapping("/reservation")
+    public String getUserReservationPage() {
+        return "reservation";
+    }
 
     @GetMapping("/admin")
     public String getMainPage() {
